Merge support for .NET 6 Lambda managed runtime into mainline (#926)
This change adds support for the .NET 6 managed runtime for lambda. It is composed of the following changes:

- #911 - property refactoring (`IsNetCore` -> `IsDotNet`)
- #914 - add .NET 6 managed runtime option into the Lambda publish wizard
- #915 - refactor lambda handler validation
- #920 - support top level statement based projects when validating Lambda handlers
  - This change updates the Lambda deploy wizard to allow handlers that are not in the "<assembly>::<type>::<method>" format for Lambda functions using top-level statements. To do this, the Toolkit detects if the project is an executable. If so, it only expects the Handler to be non-blank for .NET based managed runtimes.
- #921 - Reference the Updated blueprints collection (new version: v9 - .NET 6 related Lambda/Serverless templates)
  - The blueprints were added in #916

-----

* Preparations to Define .NET 6 runtime (#911)

This change is being done in advance of changes that will be made to the Lambda deployment wizard:
- rename property from IsNetCore to IsDotNet
- change IsDotNet from heuristic to explicit declaration

* Define .NET 6 managed runtime, start adding support into deployment wizard (#914)

This change adds the .NET 6 managed runtime option into the Lambda publish wizard.

* Move AllRequiredFieldsAreSet logic from Control to ViewModel (#915)

This change moves some Lambda validation logic around, in preparation for a future change that will update the logic on a conditional basis. The code that moved was left as-is.

The tests relating to this logic were isolated from tests that exercise the control, which is good because those tests no longer need a UI thread. As a result, `SetValidZipDeployValues` now calls `CreateDotNetHandler`, but the test code that moved is otherwise the same. The UploadFunctionViewModel_IsValidConfigurationTests gained its own Fixture, so that we only write sample files to a temp location once for all of the tests in the class, instead of once per test.

* Add support for top level statements in Lambda Handler validation (#920)

This change updates the Lambda deploy wizard to allow handlers that are not in the "<assembly>::<type>::<method>" format for Lambda functions using top-level statements. To do this, the Toolkit detects if the project is an executable. If so, it only expects the Handler to be non-blank for .NET based managed runtimes.

The help text has been updated when the project is an executable, and a .NET managed runtime is selected.

To support this, I added an overload into the wizard property code, to make it easier to retrieve a value, and use a specific default value if the property does not exist.

This change was tested with C# and F# projects (normal functions, custom runtime functions, and functions using top-level statements)

* Update the Toolkit to use the new v9 blueprint set (#921)

This change updates the Toolkit to use the new collection of blueprints, which contain Lambda functions that target .NET 6 in VS 2022. The blueprints were added in #916