Ignore empty keys in KeyValue dialog in Publish to AWS feature (#932)
If a user includes an empty/whitespace key name as an Environment Variable when publishing to Elastic Beanstalk via the Publish to AWS feature, the actual publication process will fail. This change ignores any key/value pairs where the key is empty/whitespace in either the key/value or batch text editor. Only key/value pairs with valid keys are persisted and appear in the Edit Settings UI.