Refactor Credentials Id grouping (#933)
This change centralizes how credentials identifiers are grouped for presentation in credentials profile selection UIs. This change is being made in advance of an ICredentialIdentifier selector, which is similar to the AccountViewModel selector, and will leverage the code centralized in this change.

The AWS Explorer's credentials ComboBox has the same presentation/behavior with this change.