Create CredentialId selector, add to Credential Selection Dialog (#934)
This change adds a new reusable user control for selecting from a list of `ICredentialIdentifier` objects. It is similar to an existing control for selecting from AccountViewModel objects (and leverages the same grouping logic - #933), and is intended for use in newer code, as we move away from the older heavier models.

The change has been integrated into the credentials selection dialog (#931). The dialog also received a few styling adjustments at the same time.