Add a generic dialog for handling basic selection use-cases. (#942)
Add a SelectionDialog class for use-cases where a user should be presented with a dialog containing a list box that allows for single or multiple selection of items. Follows the existing pattern of setting properties on the dialog to initialize it before showing it and retrieving values from properties if and when the Show() method returns with true (indicating the OK button was pressed).