Use default json for profile and region when deploying Lambda/Serverless (#956)
This change updates the Lambda deploy wizard to use the credentials profile and region specified in `aws-lambda-tools-defaults.json` as the initial values when present. If the file does not exist, the fields are not present, or the fields reference invalid data, the wizard falls back to using the AWS Explorer's currently selected profile/region.

This change addresses a workflow problem identified in https://github.com/aws/aws-toolkit-visual-studio/issues/224, where it is too easy for users working with projects for different AWS accounts to publish into the wrong location.

I took the opportunity to remove the coupling between `UploadFunctionController` and the Navigator (AWS Explorer). I pushed these details up the call stack, so that UploadFunctionController is provided with the information relating to the relevant account and region. The class is still not easily testable, but this helps move things in the right direction.

This change was verified with the following scenarios:
- publish Lambda project and Serverless project
- no json file present (wizard shows AWS Explorer-selected Profile/Region)
- json file exists, without profile or region fields (wizard shows AWS Explorer-selected Profile/Region)
- json file exists, with valid profile and region fields (wizard shows Profile/Region referenced in json file)
- json file exists, with non-valid profile and region fields (wizard shows AWS Explorer-selected Profile/Region)