Fix images shown after publish to work with dark theme (#969)
This change fixes a bug where the "copy to clipboard" image would not display nicely on dark theme. This is related to #946, and the cause is that the image is loaded before it is rendered, and it then does not get the appropriate adjustments based on Theme. I updated the success/failure images in case they have the same issue.