Decouple "View ECR Repo" from IViewModel (#967)
This change sets up the "View ECR Repo" command (`ViewRepositoryController`) so that it does not depend on models that come from the AWS Explorer (`RepositoryViewModel`). This allows us to leverage the command from elsewhere in the Toolkit, to view resources from other regions/profiles without first loading that connection in the AWS Explorer.

This is accomplished by defining a new command type (`IConnectionContextCommand`). This command's `Execute` call does not accept any parameters. This parallels `IContextCommand`, whose Execute call requires an `IViewModel`. The new command type is still created and invoked on the UI thread through the use of a new class `ConnectionContextCommandExecutor` , which parallels `ContextCommandExecutor`.

Where the `IContextCommand` was invoked (in the service activators), the callback now unboxes `IViewModel` and produces the controller as needed. By pulling this unboxing out of the controller/executor, we now have a controller that is easier to re-use for other needs.

This change focuses on one command (ViewRepositoryController), but provides an example we can use to migrate other commands if we realize a need to repurpose them independently of the AWS Explorer.