Fix erratic UI Control behavior when running on secondary displays (#977)
This change fixes a problem in aws/aws-toolkit-visual-studio#233. When Visual Studio is showing on a non-primary display in a multi-display environment, UI controls in the AWS Toolkit can have erratic behavior. This can be seen in Publish to AWS screens, where ComboBoxes would flicker and could not be opened. The repro case was to use a laptop and one external HDMI monitor.

This problem was not seen in modal dialogs or ToolWindows. Only in views that are hosted within document tabs.

The cause of this problem relates to how the Toolkit shows views in document tabs. To show a control in a document tab, the Toolkit hosts it in a WinForms-based shim (HostedUserControl/ElementHost), and assigns that shim to a (VS SDK) WindowPane. When the Toolkit was created, Visual Studio was only capable of hosting WinForms controls (not WPF controls). Over time, the Toolkit introduced WPF controls, and this shim approach was a way for the Toolkit to display a mix of WinForms controls and WPF controls within document tabs. Hosting WPF content within WinForms is not ideal - it can lead to a breakdown in consistent display handling (like styling, resultion, and DPI). Today we only host WPF content in document tabs, and Visual Studio is capable of directly hosting WPF content, so having this WinForms-in-the-middle shim is no longer necessary. We are able to eliminate the problem by removing the shim.