Update the Role selection dialog to optionally filter based on service principal (#984)
This change adds an optional property to the IAM Role selection dialog, which allows us to restrict the list of roles shown in the dialog to those which reference a specific service principal in the assume role policy document.

Existing Toolkit usage of the dialog remains unchanged (the dialog will show all roles). This change will be leveraged in a future commit.