Updated Blueprints to use latest SDK and Amazon Lambda packages. (#961)
This PR updates the package references for AWS packages in the blueprint. The following commits from the Lambda repo show the actual version number changes. The main goal for pushing this out is there were bug fixes in `Amazon.Lambda.Serialization.SystemTextJson` and `Amazon.Lambda.AspNetCoreServer.Hosting` released and we want the blueprints to default to using the new versions.

I have run the build script from the lambda repo to confirm blueprints work and creating the blueprints in VS and seen the updated versions are in VS.

https://github.com/aws/aws-lambda-dotnet/commit/4502d8dc63d4cf30f5ea85cb1760fbdec9e7f8b6
https://github.com/aws/aws-lambda-dotnet/commit/621e967926f1a097d22d338e37b099e37d78188e