Add confirmation prompt before publishing (#997)
This change introduces a confirmation prompt prior to publishing in the Publish to AWS experience. This gives customers a chance to back out before starting the publish, and helps reduce the need to cancel a publish (currently there is no feature to cancel a publish once started).

The prompt can be suppressed on a per-project basis. This allows users to choose to have a quicker path to deployment once they are comfortable with the publish experience.

The suppression details are stored with the rest of the Publish to AWS settings in `%localappdata%\AWSToolkit\PublishSettings.json`, where a list was added that contains the project guids (eg from .csproj files or their parent solutions) that users have suppressed confirmation prompts for.