Add Toolkit style for Expander controls (#1012)
Visual Studio does not have a built-in style for Expander controls, and the stock control's presentation is a little clunky. This change defines a style for Expanders that supports VS Theming, and better fits into UIs where it is used. The style is defined as the default for all Expanders. Currently, there is only one Expander in the product (within the Publish to AWS experience).

To create the style, I used Visual Studio to produce the basic style definition, then I pared it down to the minimum use-case for the Toolkit, changed the toggle button's presentation to use a more compact arrow (one that is consistent with TreeView controls), and applied VS Theme coloring.