Publish metric uses base recipeId when available; records generated project (#1029)
Generated projects contain a recipe definition that is also auto-generated. The recipe contains a GUID as the recipe Id, but also contains a base recipe Id, which references the recipe that was used to produce the generated project. The GUID based recipe Id has been used in the Toolkit's metrics for publish operations, which means we're unable to tell what kind of deployment took place. This change overrides the metric's recipe id value with the base recipe id, when one is available. The change also adds a new attribute to the metric, so that we can identify how often generated projects are published.

This change leverages aws/aws-dotnet-deploy#520 from the deploy tool.