Update Toolkit to .NET Framework 4.7.2 (#1030)
This change updates the Toolkit projects to use .NET Framework 4.7.2 instead of 4.6. 4.6 was necessary when the Toolkit supported VS 2017. Now that our minimum supported version is VS 2019, and VS 2019's maximum supported .NET Framework is 4.7.2, we can update to 4.7.2.

The buildtools are on 4.5 and have not been adjusted in this change, because there are internal pipelines that use those tasks.

I've also added a GitHub Actions CI job that compiles and tests the Toolkit from VS 2022. We didn't do this previously, because there were some .NET 4.6.1 references, which caused earlier attempts to set up a 2022 CI to fail, because that framework isn't installed in the image.

I've also updated one buildtools test that was failing (buildtasks/BuildTasks.Test/CreateTestVersionTaskTests.cs).