Add temporary validation to check that the Existing VPC field is not empty when creating a new VPC Connector (#1040)
This change introduces a temporary, Toolkit-side validation around VPC Connectors. If you choose to create a new VPC Connector, the validation will check that a VPC has been selected. Work on the deploy tool to support this validation will take a while, so this change serves as a stopgap, which will allow us to start sharing the newer VPC Connector feature sooner.

To introduce this validation, I created a ConfigurationDetail class `VpcConnectorVpcConfigurationDetail`, which is only used for `Vpc` fields that are children of `VPCConnector` fields.

The validation is similar to how ListConfigurationDetail are validated in #1031. This change includes a fix, where the setting was still considered invalid, even if the setting is not visible (which happens when "create new VPC" is un-checked.