Remove "Back" button; add "Modify" (target) link (#1056)
This change updates the edit settings view in the Publish to AWS experience, to remove the back button. The "Back" button doesn't articulate that it takes users to the target selection view. The button was moved to the label that shows the current publish target, renamed to indicate that it would change the target, and changed to a link.

Moving it away from the publish button also increases the focus on what should happen next (publishing), since there is only one gesture at the bottom of the edit settings view. This also reduces the chance of accidentally clicking the wrong gesture, since the "target selection" and "publish" gestures are no longer near each other.