Improve configurationDetail loading perfomance by making fewer calls to load ValueMappings (#1061)
This change improves the performance of loading ConfigurationDetails, by removing unnecessary calls to load ValueMappings. ValueMappings are used to populate dropdown lists for a displayed ConfigurationDetail, and there are certain fields where this data will not be presented. Avoiding these loads means there is less of a delay after each setting is edited in the UI.