ConfigurationDetails and their ValueMappings are loaded together (#1062)
This change updates the main Publish to AWS ViewModel to retrieve ConfigurationDetail objects and their associated ValueMappings from a single method. This improves the code flow around loading this data and updating the UI. The change is being made in advance of another change that will consume this loading operation.