UI Improvements in Target selection view (#1050)
This change makes UI adjustments to the Publish to AWS experience. The radio buttons to choose between publishing to a new target or an existing target have been changed to tab panels that have been styled as labels. Additional UI changes were made that bring the view's presentation together:
- labels were added above the target selection and sidebar
- a border was added to the target sidebar, and padding was added to the contents within
- the heading labels were adjusted so that they all align

To accommodate this change, the main viewmodel's notion of "am I a new or existing deployment" (`IsRepublish`) was replaced with a concept that represents the operation (`TargetSelectionViewModel`, `TargetSelectionMode`). This also provided a place to store the operation's current selection, meant we could remove the "caching" code from the main viewmodel. When you switch between New/Existing targets, the UI will continue to select the target that was previously selected for that selection mode.

The new TargetSelectionViewModel classes give us a possible avenue to move more state out of the larger viewmodel, but this will be left as future work, to keep this change from growing.