Fix state handling when quickly clicking on Publish to AWS targets (#1065)
This change fixes a problem where clicking on different targets in quick succession could leave the Publish to AWS experience in a state where it was not using the settings or status from the most recently selected target.

To support this change, an abstraction around CancellationTokens was created, that allows you to cancel the current token and create a new one (`ResettableCancellationToken`). This is relevant to other cancellation token sources in PublishToAwsDocumentControl.xaml.cs, but that will be done as a follow-up change.

We want to avoid trapping `OperationCanceledException` exceptions, so portions of the code where a new token source has been added have been adjusted to only catch the deploy tool's exceptions (`ApiException`). This avoids output messages relating to cancel exceptions, which would be confusing to a user, and starts to reduce our eagerness to catch exceptions everywhere.

The tests revealed that adjusting the exception catching logic would now surface the exception produced by `ThrowIfSessionIsNotCreated`. Rather than leave that as a plain old Exception, I created a dedicated Exception class for this scenario (`NoDeploymentSessionException`).