Apply invalid values and validation messages from the server (#1085)
This change builds on updating the deploy tool to 0.46.16 (in #1084). When configuration details are loaded from the back-end, they now contain validation details, which include the validation message, and the invalid value (when the setting is not valid). The invalid value and the validation message are applied to the Toolkit's ConfigurationDetail models, which are then displayed in the UI.

This change also improves the scenario where a user selects to use an existing Role instead of a new Role. The existing role field now shows a validation if it is empty, because the Toolkit is now being provided the validation status when loading settings.

I also added a changelog for a validation added to 0.46.16 relating to checking for collisions against existing resources.