Emit metric for viewing CloudWatch Logs (#1094)
This change emits a metric when the "View CloudWatch Logs" operation is performed.

This change introduces a way for us to associate metrics with a user gesture, for metrics that leverage the "serviceType" and "source" telemetry fields. `BaseMetricSource` can be derived from on a per-plugin basis, where plugins can provide the appropriate service and relevant source definitions.