CloudWatch Metrics: Open group and stream (#1100)
This change adds metrics for opening a CloudWatch Log group and a CloudWatch Log stream.

I had to change how the log streams UI loads data, so that it defers loading until the UI is loaded. This was causing problems in the tests, because it was trying to load from partially defined mocks.