Continue if ValueMappings fail to load; Surface loading errors (#1096)
If a service isn't available in a region, ValueMapping retrieval can fail (this was discovered trying to configure an AppRunner deployment for the Canada region). When this happens, the call-stack is stopped, and there is no user feedback. With this change, we surface the error to users in the output pane, and leave the affected configuration details without ValueMappings.

I also added an attempt to add a message indicating that a service may not exist in the region of interest. This is a stopgap -- the deploy tool has a task to omit deployment targets that aren't relevant for a region.

I updated RefreshTargetConfigurationsAsync to similarly surface a message to the output panel, to reduce confusion when error scenarios occur.

By submitting this pull