Fix "AWSSDK.SSOOIDC could not be found" error when using Toolkit with SSO credentials (#1134)
This change fixes a problem introduced in 8bb61e8d044 (#1038). Users working with SSO credentials were unable to connect to their accounts, because the Toolkit was not able to automatically load the SSO related AWS assemblies. This change forces the Toolkit to load those DLLs on startup.

This fix was investigated by @floralph