Bump deploy tool to 0.50.2 (#1133)
This change updates the deploy tool to 0.50.2, giving Publish to AWS the following updates:
- https://github.com/aws/aws-dotnet-deploy/releases/tag/0.50.2
- https://github.com/aws/aws-dotnet-deploy/releases/tag/0.49.14
- https://github.com/aws/aws-dotnet-deploy/releases/tag/0.48.15

In updating the deploy tool, one of the Fargate configurations (security groups) was changed from a comma-separated string field, to a list. The Toolkit has a validation for empty lists. This change allows this configuration to have an empty list, which matches the recipe's definition.

This change also updates the troubleshooting guide URL now that https://aws.github.io/aws-dotnet-deploy/troubleshooting-guide/ no longer 404's.