Improve performance of plugin detection (#1147)
This change improves Toolkit initialization performance around how plugin activators are found and instantiated.

Previously, all types in a plugin's assembly were interrogated for plugins. This was replaced with a less expensive approach, where an assembly-level attribute `PluginActivatorType` is used to declare the plugin activators that should be instantiated on startup.

In debug mode, this reduced CPU processing times from 55 seconds to 25 seconds. The plugins are processed in parallel, so the actual time spent (and saved) is considerably less. In release mode, this appears to reduce CPU processing times from 800ms to 400ms.