Move VS SDK related tests to their own test project (#1159)
This change moves all tests relating to (or involving) the VS SDK to separate test projects that can be instantiated for each supported version of the VS SDK (currently VS 2019 and 2022). The motivation for doing this is to bring the codebase to a place where we can locally develop in VS 2019 or VS 2022, without needing to compile projects exclusively targeting a different VS version, while retaining test coverage.

To do this, 
- code that uses VS SDK types was moved out of `AWSToolkit.Tests.Common` and `AWSToolkit.Tests`. References to the VS SDK metapackage were removed from the project(s).
  - a reference to the VS SDK metapackage was added to `PublishToAws.*.Tests` as a result, because these tests relied on their dependency to `AWSToolkit.Tests.Common` to get the VS SDK references
- a new (shared) project `AWSToolkit.Tests.Common.VS.Shared` has been created. It parallels `AWSToolkit.Tests.Common` in that it is intended to house testing utility code, where the utility code relies on VS SDKs.
- a new set of test projects (`AwsToolkit.Vs.vXX.Tests` and `AwsToolkit.Vs.Tests.Shared`) has been created. Moving forward, this is where tests that use VS SDK will be added (with the exception of PublishToAws and AwsToolkitPackage, which have their own test projects).
  - `AwsToolkit.Vs.vXX.Tests` are added to the msbuild target for running the tests

CONTRIBUTING.md and designs/toolkit-architecture/README.md have been updated to mention that this is the intended pattern for tests relating to VS SDK moving forward.

A reference to `System.Runtime.CompilerServices.Unsafe` was added to `AWSToolkit.Tests.Common`. This was needed in order to prevent a runtime exception (unable to locate assembly) when running the tests from the VS developer command prompt.