Update deploy tool to 1.1.15 (#1175)
This change updates the deploy tool to the latest release (1.1.15 - https://github.com/aws/aws-dotnet-deploy/releases/tag/1.1.15). This fixes the problem de-serializing deployment settings from a Fargate target that was published with a pre-GA deploy tool, where the target was configured with multiple security groups. (https://github.com/aws/aws-toolkit-visual-studio/issues/271)

The change was run against the integration tests. I also deployed a project to Fargate using a pre GA (v47) version of the deploy tool, configured to use multiple security groups, and then tried using this new version of Publish to AWS. I was able to work with the deployment target, and see the error surfaced by the deploy tool.