Add filtered solutions per major VS Version (#1176)
This change adds filtered solutions for each of VS 2019 and 2022. This way, you can load and compile fewer projects while working within a version of Visual Studio. The intention is to reclaim some time back in the development loop. In these filtered solutions, we're able to reduce the number of projects by 22 (78 of 100 projects are loaded).