Support async initialization of AWS Explorer ToolWindow (#1178)
When ToolWindows (like the AWS Explorer) are left open in a previous session of Visual Studio, the next time the IDE is launched, the ToolWindows cause a synchronous force-load of its parent extension/packages. It isn't well documented how to set up ToolWindows to help asynchronously load the packages, but I got information from Mads, and https://github.com/microsoft/VSSDK-Extensibility-Samples/tree/master/AsyncToolWindow provides a sample implementation.

This change updates the AWS Explorer ToolWindow so that the package has a greater chance of being loaded in async fashion. The key aspect to do this is to give the ToolWindow's constructor an argument, and leverage AsyncPackage callbacks like InitializeToolWindowAsync.

While working on this change, I cleaned up some related code:
- the command associated with the View -> AWS Explorer menu was moved to a separate class ViewAwsExplorerCommand
- code relating to OpenShellWindow was made async, and unused portions of code were removed