Move deploy tool initialization/validation off the UI thread (#1183)
This change improves the Toolkit's startup performance by moving the deploy tool's installation and verification into a background thread. This was running on the UI thread, which would cause other portions of the Toolkit initialization to block while waiting for access to the UI thread. This would have been particularly noteworthy when there was a new version of the deploy tool for the Toolkit to download and install.

In development with my debug build of the Toolkit, I saw a 170ms reduction in Startup/UI thread CPU usage. The savings are more pronounced on slower hardware, containing fewer cores, and/or on systems that would download new versions of the deploy tool at a slower rate.