Optimistically load more events when filtered (#1177)
This changes the way toolkit handles querying the FilterEventsAPI for large log groups containing many log streams and events. With this change, user interaction is not required to repeatedly query the API and we optimistically keep loading till the scrollbar appears and users can scroll down to load further pages.