Update Lambda blueprints to latest AWS packages (#1171)
Update the .NET Lambda blueprint to use the latest AWS packages. The update is captured in the following PR and this is the binary (zip) version of the previous PR. The main reason for the PR is to update the preview blueprint that uses the new preview Lambda HLL the DevEx has been working on. The library got some new updates and we want to get the blueprint updates to see if there is any feedback on the changes.

https://github.com/aws/aws-lambda-dotnet/pull/1266

I have confirmed the change is good using the lambda repo script that all of the templates compile as well as manually using the templates via Visual Studio Toolkit to make the changes propagated as expected.