Add region and include call to action to file an issue when there are no existing targets (#1187)
This change introduces a call to action for users to create a GitHub issue when Publish to AWS reports that there are no existing targets, but the user thought their cloud app would be in the list. This gives us a chance to further engage with them and better understand what they were expecting to see and why.

The GitHub issue is based on a URL that is pre-filled with some basic information, and contains a prompt to encourage the user to describe their scenario.

The Publish to AWS selected region has been added into the messaging as well, in case the user thought they were querying from a different region.

To make this change, I had to provide access to the IDE/Toolkit version details, so I have added these to the Toolkit Shell Provider.