Update deploy tool to 1.2.4 (#1194)
This change updates the deploy tool to the latest release (1.2.4 - https://github.com/aws/aws-dotnet-deploy/releases/tag/1.2.4). This fixes the problem where users could not deploy to Fargate in a region where there is no default VPC (aws/aws-toolkit-visual-studio#271).

This change was run against the integration tests. I also saw that I could select Fargate as a deployment choice in a region without a default VPC (but I didn't try deploying it).