Update help link in Credentials dialog (#1193)
The add/edit credentials dialog had a link to a help page that no longer exists (https://aws.amazon.com/developers/access-keys/). The dialog has been updated to link to a page in the Toolkit User Guide instead. Also, the link has been removed, in favor of using more standard dialog help conventions. Now the dialog has a help icon in the titlebar, and a help button in the footer, which will open up the user guide's credentials page. Pressing F1 in this dialog also does the same thing.

This was accomplished by leveraging the VS SDK's DialogWindow callbacks for providing a help action. I introduced `IHelpHandler` as a way of providing help without bloating `IAWSToolkitControl` with more concerns.

I also performed a little cleanup in the RegisterAccountModel file.