Add Visual Studio Folder Browser (#1199)
The standard WinForms folder browser is old, and not very user friendly. You can't copy or paste a path directly into it, and in places where the Toolkit uses it, you have to navigate the whole system folder structure every time you open it (it doesn't "remember" your last folder selection.

This change introduces a wrapper around the VS SDK folder browser, so that the Toolkit can use the same dialog as Visual Studio. The CodeCommit related folder browsers have been changed to use the new browser. Other places in the Toolkit can be updated as follow-up changes.