Define ToolkitCredentials, integrate with ICredentialProviderFactory (#1206)
This the first in a series of changes that will introduce bearer token support into the Toolkit.

This change introduces a class `ToolkitCredentials` that can hold AWSCredentials or (in a future change) a bearer token producer. This way, existing code that works with AWSCredentials objects from the Toolkit's credentials system can be changed over time to access AWSCredentials through `ToolkitCredentials`.

The Toolkit's credential system will be updated over several changes, to return `ToolkitCredentials` instead of `AWSCredentials`. In this change, the credential factories (`ICredentialProviderFactory`) have been updated to create and return `ToolkitCredentials` objects. This is one of the innermost components of the credentials system, so AWSCredentials are directly obtained from `ToolkitCredentials`. `ICredentialProviderFactory.CreateAwsCredential` has been removed, because it is only called in one place (ConnectionManager), which has now been updated to call the successor `CreateToolkitCredentials`.

Future changes will update the Connection Manager and Credential Manager to use `ToolkitCredentials`.