Integrate ToolkitCredentials with Credential Manager and Connection Manager (#1209)
This is part of a series of changes to introduce Bearer Token support into the Toolkit.

This change builds on #1206 and updates the CredentialManager and ConnectionManager to use the new `ToolkitCredentials` class as the backing object instead of the SDK's `AWSCredentials` class. This provides the Toolkit code with a way to establish bearer token based connections (in future changes), by obtaining `ToolkitCredentials` objects.

In `ICredentialManager`, a new `GetToolkitCredentials` method has been added, which will replace `GetAwsCredentials` over time, as Toolkit code is migrated over to leverage `ToolkitCredentials`. Existing Toolkit calls to `GetAwsCredentials` have been left as-is, as this migration can take place over time.

In `IAwsConnectionManager`, `ToolkitCredentials` now represents the active credentials instead of `AWSCredentials`. Token connections will have an "AWS ID", so this has been stubbed in. Lookup and assignment of AWS ID will take place in an upcoming change. AWS ID is an additive concept - `ActiveAccountId` continues to be the AWSCredentials based way of identifying a user.