Define general Token Manager; Add token provider to ToolkitCredentials (#1210)
This is part of a series of changes to introduce Bearer Token support into the Toolkit.

This change builds on #1206 to introduce the concept of the AWS .NET SDK's Token Providers into the credentials system. `ToolkitCredentials` has been updated to hold a token provider, and a generalized SSO Token manager (`ToolkitSsoTokenManager`) has been defined in the Toolkit. This token manager will be leveraged in future changes to connect with a bearer token service. This is similar to how the Toolkit wraps (non bearer token) SSO Credentials with `SSOAWSCredentials` in order to handle the login callback.