Fix missing CloudFormation project/template icons in Solution Explorer for VS 2022 (#1220)
This change adds the missing icons for CloudFormation projects and template files within the CFN projects for VS 2022. I copied the embedded resource declarations from the VS 2019 project.

It looks like these icons could be changed to use our newer image resources, but that can be left for another time, since this is a quick fix.