Query for AWS ID when a Token Provider is chosen in AwsConnectionManager (#1223)
This change updates AwsConnectionManager to query for AWS ID when working with Token Provider based credentials. The initial implementation queries from CAWS, but this may change to use Sono in the future.

To support this change, the connection manager was updated to encapsulate how AccountId and AWS ID are retrieved (`IIdentityResolver`). This replaces the function that was used to inject an STS service client.