Filter TokenProvider based credentials from places that only work with AWSCredentials (#1216)
This change updates places in the Toolkit that only support AWSCredentials based connections (eg: not TokenProvider connections), so that they filter the credentials they work when when querying the connection manager for credentials. This way token provider entries will not appear in places like AWS Explorer.

The credentials selection dialog was updated so that calling code can optionally filter the credentials displayed based on the connection types they support. The only place this is currently used is in Publish to AWS.

This change builds on #1206 and #1210.