Instrument failure reasons when publishing Serverless Applications (#1234)
The metric serverlessapplication_deploy currently reports the Toolkit's second highest amount of failures. It looks like that deploy code already produced error codes intended for use with telemetry, so this change wires in the error codes for those failures. This change will be treated as an experiment to see if the data is useful, before making similar changes in other metrics with a higher rate of failure.

I also re-generated the toolkit's telemetry code, so that it contains generator changes added in https://github.com/aws/aws-toolkit-common/pull/372

Sample metric (reason field is new)

```
    {
      "Metadata": {
        "awsAccount": "xxxxxxxxxxxx",
        "awsRegion": "us-west-2",
        "result": "Failed",
        "reason": "AccessDenied-Forbidden-S3UploadError"
      },
      "MetricName": "serverlessapplication_deploy",
      "Unit": {
        "Value": "None"
      },
      "Value": 1.0,
      "Passive": false
    }
```