Fix sort order for Beanstalk images in legacy deployment wizard (#1238)
This change fixes the order of Beanstalk images listed in the legacy deployment wizard. Images with a version of 2.10.x were being ordered lower than 2.2.x. This meant that you wouldn't see or select that image if you were looking for the top-most image, and that the wizard would pre-select a 2.7.x image (instead of the latest one) when you are deploying a .NET Framework application.