Update Amazon.Lambda.Tools to support .NET 7 AOT (#1275)
This change pulls in the latest Amazon.Lambda.Tools that supports building .NET 7 AOT Lambda functions using docker under the cover. The internal CR for the AOT change is `CR-76405166`

There are no UI changes required for this feature. Customers would enable this feature by editing the .NET 7 Custom Runtime Lambda project to have the **PublishAot** property set to true.

```
<Project Sdk="Microsoft.NET.Sdk">
  <PropertyGroup>
    <OutputType>Exe</OutputType>
    <TargetFramework>net7.0</TargetFramework>
    <ImplicitUsings>enable</ImplicitUsings>
    <Nullable>enable</Nullable>
    <AWSProjectType>Lambda</AWSProjectType>
    <AssemblyName>bootstrap</AssemblyName>
    <!-- This property makes the build directory similar to a publish directory and helps the AWS .NET Lambda Mock Test Tool find project dependencies. -->
    <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
	  
    <PublishAot>true</PublishAot>

  </PropertyGroup>
 ...
```

A future PR will add an AOT configured blueprint to VS.

There was also a feature that went out last month to add `--resolve-s3` switch that when set to true would require the CloudFormation based deployments to not have to configure an S3 bucket and just let the tooling use a default bucket and create the bucket if it doesn't exist. The toolkit could think about remove the requirement in that wizard to not force users to have to select a bucket to remove a friction point. But that is definitely an optional feature.

This change also brings in a fix for https://github.com/aws/aws-toolkit-visual-studio/issues/290, which prevented users from deploying Lambda functions to regions that do not support function urls.